Refactor chord detection tests to omit inversion index
This macro refactoring is based on the fact that one can keep track of
the inversion index, by counting tokens.

We keep track of the number of the current inversion by passing on that
many tokens, and use macro recursion to pick off the 'head' of the
inversion list, and add one token to the tokens to be counted before
recursing.

The counting is done by creating a unit array literal and calling its
len method.

    let max = count!($($max)*);

assuming the $max repetition contains the three ident:s i1, i2, i3, this
expands to;

    let max = <[()]>::len(&[(), (), ()]);

which is calling the len method of the unit-array-type [()], with an
array literal with three units; [(), (), ()].