Mute previous chord on key change
Keep track of notes playing in this slice. As we don't really know how
much of the slice has been sent already if we need to abort it (due to a
key change) we just conservatively accumulate all note on messages in a
slice. This might cause us to send more note-offs at key change than
strictly necessary, but it should not be noticable and is very simple to
do.

Note that we still forget to flush (drop) scheduled notes and mute the
currently playing notes on exit.