Impl Sub for TimeCode
Subtraction will fail if the resulting time code would be negative (as
the bar, beat and ticks are unsigned).

However, we can borrow beats from bars, so it is possible to do

    TimeCode::new(4, 0, 0) - TimeCode::new(0, 1, 0)

as we assume 4 beats to a bar, but not

    TimeCode::new(4, 0, 0) - TimeCode::new(0, 0, 1)

as we don't assume the number of ticks to a beat.