Allow specifying unit parts
The unit parts determines the length of the silence between notes.
The duration of the silence for "n" unit parts will be 1/n.

So with 4 unit parts, the length of a 2 unit note will be 1 and 3/4
units (which is the same as 7/4 units), and the duration of the silence
will be 1/4 units.

With 8 unit parts, the length of a 1 unit note will be 7/8 units, and
the duration of the silence will be 1/8 units.